AI_Project/AI_Code_Teacher/requirements.txt
google.generativeai
streamlit
os